+++ packages/common/src/Utils/component-data-validator.js
@@ -1,10 +1,10 @@
-import { Validator } from 'jsonschema';
-
-const validator = new Validator();
-
 export const validate = (componentData, schema) => {
-  const result = validator.validate(componentData, schema);
-  if (!result.valid && result.errors) {
-    result.errors.forEach(error => console.warn('Plugin schema validation error:', error)); // eslint-disable-line no-console
+  const Validator = require('jsonschema').Validator;
+  if (process.env.NODE_ENV !== 'production') {
+    const validator = new Validator();
+    const result = validator.validate(componentData, schema);
+    if (!result.valid && result.errors) {
+      result.errors.forEach(error => console.warn('Plugin schema validation error:', error)); // eslint-disable-line no-console
+    }
   }
 };